Catchup (#1)
* aligned build:env cmd with newe webpack config files

* Changed the PrimaryColor in the sepia theme

* fixed bug causing memory issues

* fix title style in note item

* animation for new note item

* UI fixes

* Fix change request

* Fixed

* fix

* Removed the unnecessary words at sign in

* added analytics wrapper for dbStore

* wrapped dbstore in analytics wrapper

* added preference analytics

* moved preferences analytics to top level

* moved configuration into hook to stop errors on test

* updated billing ux

* moved reporting to component

* remove console.log and fix equality bug

* fix all notes folder

* fix all notes folder

* capitalize notes

* target-blank for previewer links

* Fixed the billing information

* Add icon and favicon

* fix link on replace

* fix toast item

* fix id of timer

* fix id

* fix toast item timeout

* revert static

* Fix image

* fix resources links

* fix toolbar export icon

* fix

* Implement openNew for electron

* Fix external links

* Surpress updater error message

* fix

* 0.1.0-7

* side navigator small fix

* 0.1.0-8

* FIx build

* 0.1.0-9

* Add Boost Note logo

* Update readme

* fix

* Update lockfile

* 0.1.0-10

* production env and routes

* fixed open external ref

* 0.1.0-11

* fix style side nav

* remove prettier space

* Style signout button

* 0.1.0-12

* fix text align left

* Set artifactName

* fixed external links in billing

* remove sign in button when logged in

* Small fix on about page

* Fix logo link for electron

* Fix images

* Rewrite src for electron app only

* fix download app links

* fix link

* fix component

* fix app link in about page

* added publicPath for non-electron build

* formatting

* 0.1.0-13

* Add build script for electron app

* 0.1.0

* Update readme

* Update

* Fix word

* 0.1.1

* removed dummy break

* Edit CONTRIBUTING.MD

* script and imports

* add icons and sideNavigator implementation

* Changed icons

* added vertical align-middle

* fix icons PR

* Fix icons

* Fix icons

* Fix size

* fix rebase

* fix rebase

* resources to static

* fix merge

* Adjusted Colors, Sizes, and Spacing for Each Icon

* Adjusted Spacing on Sidebar

* Add setting button for storage on the side navigation

* Fix the change request

* reflect key change for other control components

* fix dev build

* add populated note doc logic

* note page modif

* fix route for non current storage

* remove console.log

* modify breadcrumbs

* fix dependency

* fix dependency

* remove fluff from utils

* active state for sidenavigator

* added bookmarked data to note

* added context menu to note list

* fixed tests with bookmarked property

* all bookmarks page

* fix unnecessary parameter for class prototype

* filter by updatedAt, createdAd, title

* Styled sidebar

* Styled select tab

* moved 'All Notes' under Storage name nav

* move notes in / to default

* disallow root (/) folder creation

* create /default folder on createStorage

* new note on right-click on folder

* createNote only on folder NoteLists

* fixed wrong storageId on sync bug

* adjust tests for removed / folder and added /default folder

* Fix sidenav

* Add NoteList

* Add remaining part

* Add Japanese

* Fix erros

* Add issue template

* Update issue template

* Fix hooks

* Fix lock file

* Fix error

* fixed error when creating a new storage becuase there were no current note

* currentNote| t

* Fixed error caused by using hooks inside of class component

* remove unnecessary translation for tuto note detail

* fixed 229 issue that checkbox dosen't check

* note create, note delete, toggle view, focus search keybinds

* after merge

* keybindings only for desktop app

* fixed app bindings

* remove key log

* toggle view to ctrl T

* fixed t -> T for use with shift with e.key

* mac fix

* fix return pretty

* Fixed Icon Colors

* Adjusted Icon Positions

* Fixed Styling of "Loading Data"

* Added SVGs for Changing View

* Added TSX for Changing View Icons

* Added New Icons

* Fixed Toolbar Icon Colors

* Tiny Styling

* Added Placeholder Color

* Use require when it exists only

* 0.1.2

* Fix require again

* Disable verify update code sign

* Add Spanish(ES) translation.

* Hide titlebar for macOS

* Update readme

* Add  solarized dark to editor themes

* Add in import

Still not working for codeblock but fixes another issue

* Revert "Add in import"

This reverts commit b8b1dffb

* Fix issues thanks to the help from Rokt33r

* Remove the console.log

* Use kebab case for theme

* Set default width and height for browesr window

* 0.1.3

* Add Chinese translation

* Fix Japanese translation

* Fix title focus issue

* Added Korean transalation.

* Fix typo errors for zh_CN

* Fix the attachments translation

* Add <ptBR.ts> to language list

* adding new ptbr files

* Refactor NotePage

* Destruct db

* Change delete note shortcut for WIndows and Linux

* Make note deletion shortcut only works when note list is focused

* Fix no storage label

* Show messageBox when delete trashedNote

* Handle the case when src is undefined

* Dashify file name of attachments

* Change to edit mode when tap createNote button

* Not create tag if create same tag already

* Fix return immediately

* Discard wrong flex style of button wrapper in NoteDetail component

* Show app version

* Create note.restore

* Show restore menu when select trashed note

* Refactor menuItems

* Update korean locale

* Show check for updates always

* Use useEffectOnce

* Fix lint errors in NoteItem

* Use toast to display error message properly

* Rollback NoteDb

* Stop creating default folder

* Fix style error in SideNavigatorItem

* Fix remove tag button label

* Fix left-padding of FoldButton

* Add all notes nav item

* Open tag list when tags clicked

* Hide global all notes item from side nav

* Try to create root folder when init

* Handle errors from notedb when new folder name is invalid

* Fix folderPathname validation to allow root folder

* Show new note button

* Exclude root folder from side nav folder list

* Revert change of init

The sync flow is really shitty. We should fix it first before fix this effect hook

* Fix tests

* 0.1.4

* Add german translation

+++ src/components/Storage/StorageEdit.tsx
@@ -1,13 +1,18 @@
 import React, { useCallback, useState, useEffect } from 'react'
-import { Section, SectionHeader } from '../PreferencesModal/styled'
-import CloudStorageSelector from './CloudStorageSelector'
-import { usePreferences } from '../../lib/preferences'
+import {
+  SectionMargin,
+  SectionHeader1,
+  RightMargin,
+  DeleteStorageButton
+} from '../PreferencesModal/styled'
 import { useDb } from '../../lib/db'
-import { CloudStorage } from '../../lib/accounts'
 import { NoteStorage } from '../../lib/db/types'
 import { useRouter } from '../../lib/router'
 import { useDebounce } from 'react-use'
-import LoginButton from '../atoms/LoginButton'
+import { useDialog, DialogIconTypes } from '../../lib/dialog'
+import { isCloudStorageData } from '../../lib/db/utils'
+import { useToast } from '../../lib/toast'
+import { useTranslation } from 'react-i18next'
 
 interface StorageEditProps {
   storage: NoteStorage
@@ -16,37 +21,47 @@ interface StorageEditProps {
 export default ({ storage }: StorageEditProps) => {
   const db = useDb()
   const router = useRouter()
-  const { preferences } = usePreferences()
+  const { t } = useTranslation()
   const [name, setName] = useState(storage.name)
+  const { messageBox } = useDialog()
+  const { pushMessage } = useToast()
 
   useEffect(() => {
     setName(storage.name)
   }, [storage])
 
-  const user = preferences['general.accounts'][0]
-
-  const linkCallback = useCallback(
-    async (cloudStorage: CloudStorage) => {
-      const success = await db.setCloudLink(storage.id, cloudStorage, user)
-      if (!success) {
-        //TODO: toast syncing failed
-      }
-    },
-    [storage.id, db, user]
-  )
-
-  const unlinkCallback = useCallback(() => {
-    db.removeCloudLink(storage.id)
-  }, [storage.id, db])
-
   const removeCallback = useCallback(() => {
-    db.removeStorage(storage.id)
-    router.push('/app')
-  }, [storage.id, db, router])
+    messageBox({
+      title: `Remove "${storage.name}" storage`,
+      message: t('storage.removeMessage'),
+      iconType: DialogIconTypes.Warning,
+      buttons: [t('storage.remove'), t('general.cancel')],
+      defaultButtonIndex: 0,
+      cancelButtonIndex: 1,
+      onClose: async (value: number | null) => {
+        if (value === 0) {
+          try {
+            await db.removeStorage(storage.id)
+            router.push('/app')
+          } catch {
+            pushMessage({
+              title: t('general.networkError'),
+              description: `An error occurred while deleting storage (id: ${storage.id})`
+            })
+          }
+        }
+      }
+    })
+  }, [storage, db, router, messageBox, pushMessage])
 
   useDebounce(
     () => {
-      db.renameStorage(storage.id, name)
+      db.renameStorage(storage.id, name).catch(() => {
+        pushMessage({
+          title: t('general.networkError'),
+          description: `An error occured while updating storage (id:${storage.id}}`
+        })
+      })
     },
     1000,
     [name]
@@ -54,49 +69,34 @@ export default ({ storage }: StorageEditProps) => {
 
   return (
     <div>
-      <Section>
-        <SectionHeader>Edit Storage</SectionHeader>
+      <SectionMargin>
+        <SectionHeader1>{t('storage.edit')}</SectionHeader1>
         <div>
-          <label>
-            Storage Name{' '}
-            <input
-              value={name}
-              onChange={e => setName(e.target.value)}
-              type='text'
-            />
-          </label>
-          <span onClick={removeCallback}>Delete Storage</span>
+          <RightMargin>
+            <label>
+              <RightMargin>{t('storage.name')}</RightMargin>
+              <input
+                value={name}
+                onChange={e => setName(e.target.value)}
+                type='text'
+              />
+            </label>
+          </RightMargin>
+          <DeleteStorageButton onClick={removeCallback}>
+            {t('storage.delete')}
+          </DeleteStorageButton>
         </div>
         <div>
-          <h2>Cloud Storage</h2>
-          {storage.cloudStorage != null ? (
-            <p>
-              Linked Storage: {storage.cloudStorage.name} (ID:
-              {storage.cloudStorage.id}){' '}
-              <span onClick={unlinkCallback}>Unlink</span>
-            </p>
-          ) : (
+          {isCloudStorageData(storage) && (
             <div>
-              {user == null && (
-                <>
-                  <p>You need to sign in to add a cloud folder</p>
-                  <LoginButton />
-                </>
-              )}
-              {user != null && (
-                <>
-                  <p>This storage is not linked yet</p>
-                  <CloudStorageSelector
-                    user={user}
-                    onSelect={linkCallback}
-                    name={storage.name}
-                  />
-                </>
-              )}
+              <p>
+                {t('storage.syncDate')}
+                {new Date(storage.cloudStorage.updatedAt).toLocaleString()}
+              </p>
             </div>
           )}
         </div>
-      </Section>
+      </SectionMargin>
     </div>
   )
 }
